🚸 Add Error handling for CRUD events

+++ app/Http/Controllers/API/EventController.php
@@ -7,7 +7,6 @@ use App\Models\Event;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Validator;
 use App\Http\Resources\EventResource;
-use Carbon\Carbon;
 
 class EventController extends Controller
 {
@@ -31,45 +30,21 @@ class EventController extends Controller
      */
     public function index()
     {
+        try {
+            $page = filter_input(INPUT_GET, "page", FILTER_SANITIZE_NUMBER_INT);
+            $range = filter_input(INPUT_GET, "range", FILTER_SANITIZE_NUMBER_INT);
 
-        $page = filter_input(INPUT_GET, "page", FILTER_SANITIZE_NUMBER_INT);
-        $range = filter_input(INPUT_GET, "range", FILTER_SANITIZE_NUMBER_INT);
+            $start_id = $range * $page;
+            $end_id = $start_id + $range + 1;
 
-        $start_id = $range * $page;
-        $end_id = $start_id + $range + 1;
-
-        $status = filter_input(INPUT_GET, "status", FILTER_SANITIZE_STRING);
-
-        if ($status === 'incoming') {
-            $events = (is_null($page) || is_null($range)) ?
-                Event::where('end_date', '>', Carbon::now())->get() :
-                Event::where('end_date', '>', Carbon::now())->get()->where('id', '>', $start_id)->where('id', '<', $end_id)->get();
+            $events = (is_null($page) || is_null($range))
+                ? Event::all()
+                : Event::where('id', '>', $start_id)->where('id', '<', $end_id)->get();
 
             return response([ 'events' => EventResource::collection($events), 'message' => 'Retrieved successfully'], 200);
+        } catch (\Exception $e) {
+            return response(['error' => $e ? $e : 'An error has occurred'], 500);
         }
-
-        if ($status === 'archived') {
-            $events = (is_null($page) || is_null($range)) ?
-                Event::where('end_date', '<', Carbon::now())->get() :
-                Event::where('end_date', '<', Carbon::now())->get()->where('id', '>', $start_id)->where('id', '<', $end_id)->get();
-
-            return response([ 'events' => EventResource::collection($events), 'message' => 'Retrieved successfully'], 200);
-        }
-
-        $events = (is_null($page) || is_null($range)) ?
-            Event::all() :
-            Event::where('id', '>', $start_id)->where('id', '<', $end_id)->get();
-
-        return response([ 'events' => EventResource::collection($events), 'message' => 'Retrieved successfully'], 200);
-    }
-
-    /**
-     * Return next event.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function next() {
-        return Event::where('end_date', '>', Carbon::now())->orderBy('end_date', 'asc')->first();
     }
 
     /**
@@ -80,20 +55,37 @@ class EventController extends Controller
      */
     public function store(Request $request)
     {
-        $data = $request->all();
+        try {
+            $data = $request->all();
 
-        $validator = Validator::make($data, [
-            'title' => 'required|string|max:255',
-            'description' => 'nullable|string|max:255',
-        ]);
+            $validator = Validator::make($data, [
+                'title' => 'required|string|max:255',
+                'description' => 'nullable|string',
+                'media' => 'nullable|file',
+            ]);
 
-        if ($validator->fails()) {
-            return response(['error' => $validator->errors(), 'Validation Error']);
+            if ($validator->fails()) {
+                return response(['errors' => $validator->messages()->get('*')], 400);
+            }
+
+            if ($request->file('media')) {
+                $uploadedFile = $request->file('media');
+                $ext = $uploadedFile->extension();
+
+                $filename = str_replace('.'.$ext, '', $uploadedFile->getClientOriginalName());
+                $date = date('Y-m-d_H-i-s');
+                $completeFilename = "{$date}_{$filename}.{$ext}";
+
+                $test = $uploadedFile->storeAs('/uploadedFiles', $completeFilename, ['disk' => 'public']);
+                $data['media'] = "{$request->getSchemeAndHttpHost()}/storage/uploadedFiles/{$completeFilename}";
+            }
+
+            $event = Event::create([...$data, "user_id" => $request->user()->id]);
+
+            return response(['event' => new EventResource($event), 'message' => 'Created successfully'], 201);
+        } catch (\Exception $e) {
+            return response(['error' => $e ? $e : 'An error has occurred'], 500);
         }
-
-        $event = Event::create($data);
-
-        return response(['event' => new EventResource($event), 'message' => 'Created successfully'], 201);
     }
 
     /**
@@ -102,9 +94,19 @@ class EventController extends Controller
      * @param  \App\Models\Event  $event
      * @return \Illuminate\Http\Response
      */
-    public function show(Event $event)
+    public function show($event_id)
     {
-        return response(['event' => new EventResource($event), 'message' => 'Retrieved successfully'], 200);
+        try {
+            $event = Event::find($event_id);
+
+            if ($event === null) {
+                return response(['message' => 'Event not found'], 404);
+            }
+
+            return response(['event' => new EventResource($event), 'message' => 'Retrieved successfully'], 200);
+        } catch (\Exception $e) {
+            return response(['error' => $e ? $e : 'An error has occurred'], 500);
+        }
     }
 
     /**
@@ -114,22 +116,44 @@ class EventController extends Controller
      * @param  \App\Models\Event  $event
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, Event $event)
+    public function update(Request $request, $event_id)
     {
-        $data = $request->all();
+        try {
+            $event = Event::find($event_id);
 
-        $validator = Validator::make($data, [
-            'title' => 'required|string|max:255',
-            'description' => 'nullable|string|max:255',
-        ]);
+            if ($event === null) {
+                return response(['message' => 'Event not found'], 404);
+            }
+            $data = $request->all();
 
-        if ($validator->fails()) {
-            return response(['error' => $validator->errors(), 'Validation Error']);
+            $validator = Validator::make($data, [
+                'title' => 'string|max:255',
+                'description' => 'nullable|string',
+                'media' => 'nullable|file',
+            ]);
+
+            if ($validator->fails()) {
+                return response(['errors' => $validator->messages()->get('*')], 400);
+            }
+
+            if ($request->file('media')) {
+                $uploadedFile = $request->file('media');
+                $ext = $uploadedFile->extension();
+
+                $filename = str_replace('.'.$ext, '', $uploadedFile->getClientOriginalName());
+                $date = date('Y-m-d_H-i-s');
+                $completeFilename = "{$date}_{$filename}.{$ext}";
+
+                $test = $uploadedFile->storeAs('/uploadedFiles', $completeFilename, ['disk' => 'public']);
+                $data['media'] = "{$request->getSchemeAndHttpHost()}/storage/uploadedFiles/{$completeFilename}";
+            }
+
+            $event->update([...$data, "user_id" => $request->user()->id]);
+
+            return response(['event' => new EventResource($event), 'message' => 'Update successfully'], 200);
+        } catch (\Exception $e) {
+            return response(['error' => $e ? $e : 'An error has occurred'], 500);
         }
-
-        $event->update($data);
-
-        return response(['event' => new EventResource($event), 'message' => 'Update successfully'], 200);
     }
 
     /**
@@ -138,10 +162,19 @@ class EventController extends Controller
      * @param  \App\Models\Event  $event
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Event $event)
+    public function destroy($event_id)
     {
-        $event->delete();
+        try {
+            $event = Event::find($event_id);
 
-        return response(['message' => 'Deleted']);
+            if ($event === null) {
+                return response(['message' => 'Event not found'], 404);
+            }
+            $event->delete();
+
+            return response(['event' => $event]);
+        } catch (\Exception $e) {
+            return response(['error' => $e ? $e : 'An error has occurred'], 500);
+        }
     }
 }
